# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tool {
  id: number;
  name: string;
  description: string;
  link: string;
  orderNum: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateToolBody {
  name: string;
  description?: string;
  link: string;
  orderNum?: number;
  isActive?: boolean;
}

export interface UpdateToolBody {
  name?: string;
  description?: string;
  link?: string;
  orderNum?: number;
  isActive?: boolean;
}

export interface ToolStatusResult {
  id: number;
  online: boolean;
  responseTime?: number;
  checkedAt: string;
}

export interface SiteSettings {
  siteTitle: string;
  siteSubtitle: string;
  discordLink: string;
  discordButtonText: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface ErrorResponse {
  error: string;
}