# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tools
 */
export const ListToolsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  link: zod.string(),
  orderNum: zod.number(),
  isActive: zod.boolean(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListToolsResponse = zod.array(ListToolsResponseItem);

/**
 * @summary Create a tool (admin only)
 */
export const CreateToolBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  link: zod.string(),
  orderNum: zod.number().optional(),
  isActive: zod.boolean().optional(),
});

/**
 * @summary Update a tool (admin only)
 */
export const UpdateToolParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateToolBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  link: zod.string().optional(),
  orderNum: zod.number().optional(),
  isActive: zod.boolean().optional(),
});

export const UpdateToolResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  link: zod.string(),
  orderNum: zod.number(),
  isActive: zod.boolean(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a tool (admin only)
 */
export const DeleteToolParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Check live status of a tool URL
 */
export const CheckToolStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const CheckToolStatusResponse = zod.object({
  id: zod.number(),
  online: zod.boolean(),
  responseTime: zod.number().optional(),
  checkedAt: zod.date(),
});

/**
 * @summary Get site settings
 */
export const GetSettingsResponse = zod.object({
  siteTitle: zod.string(),
  siteSubtitle: zod.string(),
  discordLink: zod.string(),
  discordButtonText: zod.string(),
});

/**
 * @summary Update site settings (admin only)
 */
export const UpdateSettingsBody = zod.object({
  siteTitle: zod.string(),
  siteSubtitle: zod.string(),
  discordLink: zod.string(),
  discordButtonText: zod.string(),
});

export const UpdateSettingsResponse = zod.object({
  siteTitle: zod.string(),
  siteSubtitle: zod.string(),
  discordLink: zod.string(),
  discordButtonText: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
});