# lib/api-zod/src/generated/types/createToolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateToolBody {
  name: string;
  description?: string;
  link: string;
  orderNum?: number;
  isActive?: boolean;
}