# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  siteTitle: string;
  siteSubtitle: string;
  discordLink: string;
  discordButtonText: string;
}