# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: number;
  name: string;
  description: string;
  link: string;
  orderNum: number;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}