# lib/api-zod/src/generated/types/toolStatusResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolStatusResult {
  id: number;
  online: boolean;
  responseTime?: number;
  checkedAt: Date;
}