# lib/api-zod/src/generated/types/updateToolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lenzo Beam Central API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateToolBody {
  name?: string;
  description?: string;
  link?: string;
  orderNum?: number;
  isActive?: boolean;
}